// src/app/models/ratings-model.ts
// Generated by https://quicktype.io

export interface RatingsResponse {
  ratings: Rating[];
  meta: Meta;
  links: Link[];
}

export interface Link {
  rel: Rel;
  href: string;
}

export type Rel = 'self';
// export enum Rel {
//   Self = 'self',
// }

export interface Meta {
  dataSource: string;
  extractDate: string;
  itemCount: number;
  returncode: string;
  totalCount: number;
  totalPages: number;
  pageSize: number;
  pageNumber: number;
}

export interface Rating {
  ratingId: number;
  ratingName: string;
  ratingKey: string;
  ratingKeyName: string;
  schemeTypeId: number;
  links: Link[];
}

export const Ratings: Rating[] = [
  // {
  //   "ratings": [
  {
    ratingId: 12,
    ratingName: '5',
    ratingKey: 'fhrs_5_en-gb',
    ratingKeyName: '5',
    schemeTypeId: 1,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/12',
      },
    ],
  },
  {
    ratingId: 11,
    ratingName: '4',
    ratingKey: 'fhrs_4_en-gb',
    ratingKeyName: '4',
    schemeTypeId: 1,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/11',
      },
    ],
  },
  {
    ratingId: 10,
    ratingName: '3',
    ratingKey: 'fhrs_3_en-gb',
    ratingKeyName: '3',
    schemeTypeId: 1,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/10',
      },
    ],
  },
  {
    ratingId: 9,
    ratingName: '2',
    ratingKey: 'fhrs_2_en-gb',
    ratingKeyName: '2',
    schemeTypeId: 1,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/9',
      },
    ],
  },
  {
    ratingId: 8,
    ratingName: '1',
    ratingKey: 'fhrs_1_en-gb',
    ratingKeyName: '1',
    schemeTypeId: 1,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/8',
      },
    ],
  },
  {
    ratingId: 7,
    ratingName: '0',
    ratingKey: 'fhrs_0_en-gb',
    ratingKeyName: '0',
    schemeTypeId: 1,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/7',
      },
    ],
  },
  {
    ratingId: 6,
    ratingName: 'Pass',
    ratingKey: 'fhis_pass_en-gb',
    ratingKeyName: 'Pass',
    schemeTypeId: 2,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/6',
      },
    ],
  },
  {
    ratingId: 4,
    ratingName: 'Improvement Required',
    ratingKey: 'fhis_improvement_required_en-gb',
    ratingKeyName: 'ImprovementRequired',
    schemeTypeId: 2,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/4',
      },
    ],
  },
  {
    ratingId: 2,
    ratingName: 'Awaiting Publication',
    ratingKey: 'fhis_awaiting_publication_en-gb',
    ratingKeyName: 'AwaitingPublication',
    schemeTypeId: 2,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/2',
      },
    ],
  },
  {
    ratingId: 1,
    ratingName: 'Awaiting Inspection',
    ratingKey: 'fhis_awaiting_inspection_en-gb',
    ratingKeyName: 'AwaitingInspection',
    schemeTypeId: 2,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/1',
      },
    ],
  },
  {
    ratingId: 3,
    ratingName: 'Exempt',
    ratingKey: 'fhis_exempt_en-gb',
    ratingKeyName: 'Exempt',
    schemeTypeId: 2,
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/ratings/3',
      },
    ],
  },
  //   ],
  //   "meta": {
  //     "dataSource": "API",
  //     "extractDate": "2021-07-20T11:22:54.8997045+01:00",
  //     "itemCount": 11,
  //     "returncode": "OK",
  //     "totalCount": 11,
  //     "totalPages": 1,
  //     "pageSize": 11,
  //     "pageNumber": 1
  //   },
  //   "links": [
  //     {
  //       "rel": "self",
  //       "href": "http://api.ratings.food.gov.uk/ratings"
  //     }
  //   ]
  // }
];