// src/app/models/score-descriptors.model.ts
export interface Link {
  rel: string;
  href: string;
}

export interface Meta {
  dataSource: string;
  extractDate: string;
  itemCount: number;
  returncode: string;
  totalCount: number;
  totalPages: number;
  pageSize: number;
  pageNumber: number;
}

export interface ScoreDescriptor {
  Id: number;
  ScoreCategory: string;
  Score: number;
  Description: string;
  links: Link[];
}

export const ScoreDescriptors: ScoreDescriptor[] = [
  // {
  //   "scoreDescriptors": [
  {
    Id: 1,
    ScoreCategory: 'Confidence',
    Score: 0,
    Description: 'Very good',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/1',
      },
    ],
  },
  {
    Id: 2,
    ScoreCategory: 'Confidence',
    Score: 5,
    Description: 'Good',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/2',
      },
    ],
  },
  {
    Id: 3,
    ScoreCategory: 'Confidence',
    Score: 10,
    Description: 'Generally satisfactory',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/3',
      },
    ],
  },
  {
    Id: 4,
    ScoreCategory: 'Confidence',
    Score: 20,
    Description: 'Major improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/4',
      },
    ],
  },
  {
    Id: 5,
    ScoreCategory: 'Confidence',
    Score: 30,
    Description: 'Urgent improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/5',
      },
    ],
  },
  {
    Id: 6,
    ScoreCategory: 'Hygiene',
    Score: 0,
    Description: 'Very good',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/6',
      },
    ],
  },
  {
    Id: 7,
    ScoreCategory: 'Hygiene',
    Score: 5,
    Description: 'Good',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/7',
      },
    ],
  },
  {
    Id: 8,
    ScoreCategory: 'Hygiene',
    Score: 10,
    Description: 'Generally satisfactory',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/8',
      },
    ],
  },
  {
    Id: 9,
    ScoreCategory: 'Hygiene',
    Score: 15,
    Description: 'Improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/9',
      },
    ],
  },
  {
    Id: 10,
    ScoreCategory: 'Hygiene',
    Score: 20,
    Description: 'Major improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/10',
      },
    ],
  },
  {
    Id: 11,
    ScoreCategory: 'Hygiene',
    Score: 25,
    Description: 'Urgent improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/11',
      },
    ],
  },
  {
    Id: 12,
    ScoreCategory: 'Structural',
    Score: 0,
    Description: 'Very good',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/12',
      },
    ],
  },
  {
    Id: 13,
    ScoreCategory: 'Structural',
    Score: 5,
    Description: 'Good',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/13',
      },
    ],
  },
  {
    Id: 14,
    ScoreCategory: 'Structural',
    Score: 10,
    Description: 'Generally satisfactory',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/14',
      },
    ],
  },
  {
    Id: 15,
    ScoreCategory: 'Structural',
    Score: 15,
    Description: 'Improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/15',
      },
    ],
  },
  {
    Id: 16,
    ScoreCategory: 'Structural',
    Score: 20,
    Description: 'Major improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/16',
      },
    ],
  },
  {
    Id: 17,
    ScoreCategory: 'Structural',
    Score: 25,
    Description: 'Urgent improvement necessary',
    links: [
      {
        rel: 'self',
        href: 'http://api.ratings.food.gov.uk/scoredescriptors/17',
      },
    ],
  },
  //   ],
  //   "meta": {
  //     "dataSource": "API",
  //     "extractDate": "2021-07-20T17:31:08.7535487+01:00",
  //     "itemCount": 17,
  //     "returncode": "OK",
  //     "totalCount": 17,
  //     "totalPages": 1,
  //     "pageSize": 17,
  //     "pageNumber": 1
  //   },
  //   "links": [
  //     {
  //       "rel": "self",
  //       "href": "http://api.ratings.food.gov.uk/scoredescriptors"
  //     }
  //   ]
  // }
];